togli lo spazio

+++ FirstProgram/Program.cs
@@ -13,7 +13,17 @@
     int index = random.Next(nomi.Length); // genera un numero random tra 0 e la lunghezza dell'array
     Console.WriteLine(nomi[index]); // stampa il nome corrispondente all'indice generato casualmente
     string path2 = @"test2.txt";
-    File.Create(path2).Close();
-    File.AppendAllText(path2, $"{nomi[index]} \n");
+    if (!File.Exists(path2)) // controlla se il file esiste
+    {
+      File.Create(path2).Close();
+    }
+    if (!File.ReadAllLines(path2).Contains(nomi[index])) // controlla se il nome sorteggiato è già presente nel file
+    {
+      File.AppendAllText(path2, $"{nomi[index]}\n");
+    }
+    else
+    {
+      Console.WriteLine("il nome è già presente nel file");
+    }
   }
 }
